refactor(sdk): consolidate vega catalog table/scan APIs into vega module
Move listTablesWithColumns, scanMetadata, and scanDatasourceMetadata from
the data-connection module to the vega module, where they belong — these
endpoints talk exclusively to vega-backend. datasources.ts/datasources.py
keep thin re-exports/delegations so existing callers don't break.
