+++ packages/python/src/kweaver/resources/vega/catalogs.py
@@ -1,7 +1,7 @@
 """VegaCatalogsResource — catalog CRUD + discover/health operations."""
 from __future__ import annotations
 from typing import Any, TYPE_CHECKING
-from kweaver.types import VegaCatalog, VegaResource
+from kweaver.types import Column, Table, VegaCatalog, VegaResource
 
 if TYPE_CHECKING:
     from kweaver._http import HttpClient
@@ -65,3 +65,90 @@ class VegaCatalogsResource:
         data = self._http.get(f"{self._BASE}/{id}/resources", params=params)
         entries = data.get("entries", data.get("data", [])) if isinstance(data, dict) else data
         return [VegaResource(**e) for e in entries]
+
+    # ── Scan & table listing ──────────────────────────────────────────────
+
+    def scan_metadata(self, id: str) -> dict:
+        """Trigger a metadata scan (discover) for a vega catalog, wait for it.
+
+        ``id`` is a vega catalog id (e.g. ``d7nicrcjto2s73d9g67g``).
+        Returns the discover endpoint's response as a dict.
+        """
+        return self.discover(id, wait=True)
+
+    def list_tables(
+        self,
+        id: str,
+        *,
+        keyword: str | None = None,
+        limit: int | None = None,
+        offset: int | None = None,
+        auto_scan: bool = True,
+    ) -> list[Table]:
+        """List tables with columns from a vega catalog.
+
+        Two-stage fetch:
+          1. ``GET /catalogs/{id}/resources?category=table`` — list summaries
+          2. For each resource: ``GET /resources/{rid}`` — pull
+             ``source_metadata.columns``
+
+        If no table resources exist and ``auto_scan=True``, triggers a
+        discover and retries once. ``keyword`` filters summaries client-side
+        before the per-resource detail fetches, narrowing N+1 to k+1.
+        """
+        def _list_summaries_raw() -> list[dict[str, Any]]:
+            params: dict[str, Any] = {"category": "table"}
+            if limit is not None:
+                params["limit"] = limit
+            if offset is not None:
+                params["offset"] = offset
+            data = self._http.get(f"{self._BASE}/{id}/resources", params=params)
+            return (
+                data
+                if isinstance(data, list)
+                else (data.get("entries") or data.get("data") or [])
+            )
+
+        summaries = _list_summaries_raw()
+        if not summaries and auto_scan:
+            self.scan_metadata(id)
+            summaries = _list_summaries_raw()
+
+        if keyword:
+            k = keyword.lower()
+            summaries = [it for it in summaries if k in str(it.get("name", "")).lower()]
+
+        tables: list[Table] = []
+        for s in summaries:
+            rid = s.get("id", "")
+            if not rid:
+                continue
+            try:
+                detail_raw = self._http.get(f"/api/vega-backend/v1/resources/{rid}")
+            except Exception as exc:
+                raise RuntimeError(
+                    f"vega resource {rid} fetch failed: {exc}"
+                ) from exc
+
+            detail = detail_raw
+            if isinstance(detail_raw, dict):
+                entries = detail_raw.get("entries") or detail_raw.get("data")
+                if isinstance(entries, list) and entries:
+                    detail = entries[0]
+            if not isinstance(detail, dict):
+                continue
+            columns_raw = (detail.get("source_metadata") or {}).get("columns") or []
+            tables.append(
+                Table(
+                    name=detail.get("name", s.get("name", "")),
+                    columns=[
+                        Column(
+                            name=c.get("name", c.get("field_name", "")),
+                            type=c.get("type", c.get("field_type", "varchar")),
+                            comment=c.get("description") or c.get("comment"),
+                        )
+                        for c in columns_raw
+                    ],
+                )
+            )
+        return tables
