+++ packages/typescript/src/api/vega.ts
@@ -890,3 +890,229 @@ export async function listAllVegaResources(options: ListAllVegaResourcesOptions)
   });
 }
 
+// ── Catalog Table Listing & Scan ─────────────────────────────────────────────
+//
+// These functions operate on vega catalog ids (short slugs like
+// `d7nicrcjto2s73d9g67g`), never legacy data-connection datasource UUIDs.
+// They were originally in datasources.ts but belong here because they talk
+// exclusively to vega-backend.
+
+export interface ListTablesWithColumnsOptions {
+  baseUrl: string;
+  accessToken: string;
+  /** A vega catalog id, not a legacy data-connection datasource UUID. */
+  id: string;
+  keyword?: string;
+  limit?: number;
+  offset?: number;
+  businessDomain?: string;
+  autoScan?: boolean;
+}
+
+interface VegaResourceListEntry {
+  id: string;
+  name: string;
+  category?: string;
+}
+
+interface VegaResourceDetail {
+  id: string;
+  name: string;
+  source_metadata?: { columns?: Array<Record<string, unknown>> };
+  primary_keys?: string[];
+  [key: string]: unknown;
+}
+
+/**
+ * List tables with column details from a vega catalog.
+ *
+ * Two-stage fetch:
+ *   1. GET /api/vega-backend/v1/catalogs/{id}/resources?category=table — list summaries
+ *   2. For each resource: GET /api/vega-backend/v1/resources/{rid} — pull source_metadata.columns
+ *
+ * If the catalog has no resources and `autoScan=true`, triggers a discover and
+ * retries the list once. The optional `keyword` filters summaries client-side
+ * before the per-resource detail fetches — useful to keep N+1 down to k+1.
+ *
+ * `id` is a vega catalog id.
+ */
+export async function listTablesWithColumns(
+  options: ListTablesWithColumnsOptions,
+): Promise<string> {
+  const {
+    baseUrl,
+    accessToken,
+    id,
+    keyword,
+    limit,
+    offset,
+    businessDomain = "bd_public",
+    autoScan = true,
+  } = options;
+
+  async function listResourceSummaries(): Promise<VegaResourceListEntry[]> {
+    const body = await listVegaCatalogResources({
+      baseUrl,
+      accessToken,
+      id,
+      category: "table",
+      limit,
+      offset,
+      businessDomain,
+    });
+    const parsed = JSON.parse(body) as
+      | Array<VegaResourceListEntry>
+      | { entries?: VegaResourceListEntry[]; data?: VegaResourceListEntry[] };
+    return Array.isArray(parsed) ? parsed : (parsed.entries ?? parsed.data ?? []);
+  }
+
+  let summaries = await listResourceSummaries();
+  if (summaries.length === 0 && autoScan) {
+    await scanMetadata({ baseUrl, accessToken, id, businessDomain });
+    summaries = await listResourceSummaries();
+  }
+
+  // Keyword filter applied after autoScan guard: if the catalog has tables but
+  // keyword matches none, we must NOT trigger a redundant discover.
+  if (keyword) {
+    const k = keyword.toLowerCase();
+    summaries = summaries.filter((it) => it.name.toLowerCase().includes(k));
+  }
+
+  const details = await Promise.all(
+    summaries.map(async (s) => {
+      let body: string;
+      try {
+        body = await getVegaResource({
+          baseUrl,
+          accessToken,
+          id: s.id,
+          businessDomain,
+        });
+      } catch (err) {
+        const reason = err instanceof Error ? err.message : String(err);
+        throw new Error(`vega resource ${s.id} fetch failed: ${reason}`);
+      }
+      const parsed = JSON.parse(body) as
+        | VegaResourceDetail
+        | { entries?: VegaResourceDetail[]; data?: VegaResourceDetail[] };
+      if (Array.isArray((parsed as { entries?: unknown }).entries)) {
+        const arr = (parsed as { entries: VegaResourceDetail[] }).entries;
+        if (arr.length === 0) {
+          throw new Error(`vega resource ${s.id} returned empty entries`);
+        }
+        return arr[0]!;
+      }
+      if (Array.isArray((parsed as { data?: unknown }).data)) {
+        const arr = (parsed as { data: VegaResourceDetail[] }).data;
+        if (arr.length === 0) {
+          throw new Error(`vega resource ${s.id} returned empty data`);
+        }
+        return arr[0]!;
+      }
+      return parsed as VegaResourceDetail;
+    }),
+  );
+
+  const tables: Array<{
+    name: string;
+    columns: Array<{ name: string; type: string; comment?: string; isPrimaryKey?: boolean }>;
+    primaryKeys?: string[];
+  }> = [];
+
+  for (const d of details) {
+    const columnsRaw = (d.source_metadata?.columns ?? []) as Array<Record<string, unknown>>;
+    const tablePkArray = extractPrimaryKeys(d as unknown as Record<string, unknown>);
+    const columns = columnsRaw.map((c) => {
+      const name = String(c.name ?? c.field_name ?? "");
+      const flagged = isColumnPrimaryKey(c) || tablePkArray.includes(name);
+      return {
+        name,
+        type: String(c.type ?? c.field_type ?? "varchar"),
+        comment: typeof c.description === "string"
+          ? c.description
+          : (typeof c.comment === "string" ? c.comment : undefined),
+        ...(flagged ? { isPrimaryKey: true } : {}),
+      };
+    });
+    const synthesizedPks = tablePkArray.length > 0
+      ? tablePkArray
+      : columns.filter((c) => c.isPrimaryKey).map((c) => c.name);
+
+    tables.push({
+      name: d.name,
+      columns,
+      ...(synthesizedPks.length > 0 ? { primaryKeys: synthesizedPks } : {}),
+    });
+  }
+
+  return JSON.stringify(tables);
+}
+
+// Two PK metadata shapes are recognized — both confirmed conventions:
+//   - per-column `is_primary_key: true` (data-connection metadata standard)
+//   - per-column `column_key === "PRI"` (MySQL INFORMATION_SCHEMA pass-through)
+//   - table-level `primary_keys: string[]` (composite-PK carrier)
+// Other plausible spellings (camelCase, singular keys, SQLite `pk` integer) are
+// intentionally NOT recognized here — adding them speculatively risks false
+// matches and creates code paths the test suite can't pin down. Extend only when
+// a real backend response demonstrates the need.
+function isColumnPrimaryKey(col: Record<string, unknown>): boolean {
+  if (col.is_primary_key === true) return true;
+  if (typeof col.column_key === "string" && col.column_key.toUpperCase() === "PRI") return true;
+  return false;
+}
+
+function extractPrimaryKeys(table: Record<string, unknown>): string[] {
+  const arr = table.primary_keys;
+  if (Array.isArray(arr)) {
+    return arr.filter((x): x is string => typeof x === "string");
+  }
+  return [];
+}
+
+export interface ScanMetadataOptions {
+  baseUrl: string;
+  accessToken: string;
+  id: string;
+  /** Retained for signature compatibility; ignored — vega catalog already knows its connector_type. */
+  dsType?: string;
+  businessDomain?: string;
+}
+
+/**
+ * Trigger a metadata scan for a vega catalog and wait for completion.
+ * `id` is a vega catalog id (e.g. `d7nicrcjto2s73d9g67g`), not a legacy
+ * data-connection datasource UUID.
+ */
+export async function scanMetadata(options: ScanMetadataOptions): Promise<string> {
+  const { baseUrl, accessToken, id, businessDomain = "bd_public" } = options;
+  return discoverVegaCatalog({
+    baseUrl,
+    accessToken,
+    id,
+    wait: true,
+    businessDomain,
+  });
+}
+
+export interface ScanDatasourceMetadataOptions {
+  baseUrl: string;
+  accessToken: string;
+  id: string;
+  businessDomain?: string;
+}
+
+/**
+ * Trigger a metadata scan and wait for completion. `id` is a vega catalog id.
+ *
+ * @deprecated Use {@link scanMetadata} directly. This wrapper exists only for
+ * backward compatibility with callers that used the old data-connection-based
+ * `scanDatasourceMetadata` signature.
+ */
+export async function scanDatasourceMetadata(
+  options: ScanDatasourceMetadataOptions,
+): Promise<string> {
+  return scanMetadata(options);
+}
+
