+++ packages/typescript/test/datasources-pk-propagation.test.ts
@@ -1,7 +1,7 @@
 import test from "node:test";
 import assert from "node:assert/strict";
 
-import { listTablesWithColumns } from "../src/api/datasources.js";
+import { listTablesWithColumns } from "../src/api/vega.js";
 
 const originalFetch = globalThis.fetch;
